Use the JCS_EXT_BGRA output color space for decoding
This removes the need to allocate a second buffer for the decoded pixel
data, and may allow libjpeg-turbo to use hardware acceleration when
decoding into the the BGRA row buffer.

The access violation with gray scale images that was encountered during
the previous attempt at this appears to have been fixed with the
rewrite of the compression code to use the JCS_EXT_BGRA input color
space.
Not sure what the underlying issue was, perhaps the JPEG output color
space was not being set correctly for gray scale images.